requirements.txt
fastapi
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
idna==3.10
motor==3.6.0
passlib==1.7.4
pip==24.3.1
pyasn1==0.6.1
pydantic==2.10.4
pydantic_core==2.27.2
pymongo==4.9.2
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
setuptools==49.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0